refactor: :recycle: Updates to improve code reduction
Updates to improve code reduction
