+++ app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { List } from "@/components/list";
+import { List } from "@/components/listOld";
 import { useAppSelector } from "@/lib/redux/hooks";
 
 export default function Home() {
